Replacing duplication of template & concrete test by abstract test
