+++ sonar-html-plugin/src/test/java/org/sonar/plugins/html/checks/structure/AbstractIllegalElementCheckTest.java
@@ -1,0 +1,75 @@
+/*
+ * SonarQube HTML
+ * Copyright (C) 2010-2025 SonarSource Sàrl
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the Sonar Source-Available License Version 1, as published by SonarSource SA.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the Sonar Source-Available License for more details.
+ *
+ * You should have received a copy of the Sonar Source-Available License
+ * along with this program; if not, see https://sonarsource.com/license/ssal/
+ */
+package org.sonar.plugins.html.checks.structure;
+
+import java.io.File;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.RegisterExtension;
+import org.sonar.plugins.html.checks.CheckMessagesVerifierRule;
+import org.sonar.plugins.html.checks.TestHelper;
+import org.sonar.plugins.html.visitor.HtmlSourceCode;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+abstract class AbstractIllegalElementCheckTest {
+
+  @RegisterExtension
+  public CheckMessagesVerifierRule checkMessagesVerifier = new CheckMessagesVerifierRule();
+
+  protected abstract AbstractIllegalElementCheck createCheck();
+
+  @Test
+  void detected() {
+    assertThat(createCheck().elements).isEmpty();
+  }
+
+  @Test
+  void custom() {
+    AbstractIllegalElementCheck check = createCheck();
+    check.elements = "title,body";
+
+    HtmlSourceCode sourceCode = TestHelper.scan(new File("src/test/resources/checks/IllegalElementCheck.html"), check);
+
+    checkMessagesVerifier.verify(sourceCode.getIssues())
+        .next().atLocation(3, 2, 3, 9).withMessage("Remove this \"title\" element.");
+  }
+
+  @Test
+  void namespaceSpecificMatch() {
+    AbstractIllegalElementCheck check = createCheck();
+    check.elements = "f:inputText";
+
+    HtmlSourceCode sourceCode = TestHelper.scan(new File("src/test/resources/checks/IllegalElementTemplateCheck.html"), check);
+
+    checkMessagesVerifier.verify(sourceCode.getIssues())
+        .next().atLocation(6, 2, 6, 38).withMessage("Remove this \"f:inputText\" element.");
+  }
+
+  @Test
+  void namespaceWildcardMatch() {
+    AbstractIllegalElementCheck check = createCheck();
+    check.elements = "inputText";
+
+    HtmlSourceCode sourceCode = TestHelper.scan(new File("src/test/resources/checks/IllegalElementTemplateCheck.html"), check);
+
+    checkMessagesVerifier.verify(sourceCode.getIssues())
+        .next().atLocation(6, 2, 6, 38).withMessage("Remove this \"inputText\" element.")
+        .next().atLocation(7, 2, 7, 39).withMessage("Remove this \"inputText\" element.")
+        .next().atLocation(8, 2, 8, 39).withMessage("Remove this \"inputText\" element.");
+  }
+
+}
